+++ protocol-simulator-iec61850/src/main/java/com/alliander/osgp/simulator/protocol/iec61850/server/logicaldevices/GasFurnace.java
@@ -283,7 +283,7 @@ public class GasFurnace extends LogicalDevice {
         }
 
         if (FLOAT32_NODES.contains(node)) {
-            return this.setFixedFloat(node, fc, Integer.parseInt(value));
+            return this.setFixedFloat(node, fc, Float.parseFloat(value));
         }
 
         if (INT8_NODES.contains(node)) {
